requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cryptography==40.0.2
cssselect2==0.7.0
Deprecated==1.2.13
Django==4.1
django-crispy-forms==2.0
django-filter==23.2
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
freetype-py==2.3.0
gevent==22.10.2
greenlet==2.0.2
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
jwt==1.3.1
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
pycairo==1.23.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-bidi==0.4.2
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
redis==4.5.4
reportlab==3.6.13
requests==2.31.0
rlPyCairo==0.2.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
swagger-spec-validator==3.0.3
tinycss2==1.2.1
typing_extensions==4.6.1
tzdata==2022.7
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
xhtml2pdf==0.2.11
xmltodict==0.13.0
zope.event==4.6
zope.interface==6.0
